+++ pyccel/parser/semantic.py
@@ -4020,14 +4020,9 @@ class SemanticParser(BasicParser):
                 return If(true_section, false_section)
 
         # Visit object
-        if isinstance(rhs, FunctionCall):
-            name = rhs.funcdef
-            rhs = self._visit(rhs)
-            if isinstance(rhs, (PythonMap, PythonZip, PythonEnumerate, PythonRange)):
-                errors.report(f"{type(rhs)} cannot be saved to variables", symbol=expr, severity='fatal')
-
-        else:
-            rhs = self._visit(rhs)
+        rhs = self._visit(rhs)
+        if isinstance(rhs, (PythonMap, PythonZip, PythonEnumerate, PythonRange)):
+            errors.report(f"{type(rhs)} cannot be saved to variables", symbol=expr, severity='fatal')
 
         if isinstance(rhs, NumpyResultType):
             errors.report("Cannot assign a datatype to a variable.",